Add files via upload
This is the codes I did with more focus on the hashtags at the end. There are too many hashtags and I tried to filter the ones to exclude in excel and still there are too many. I got a bit stuck on this, because there are  many unseful ones that can lead to non sense tweets :( (but many useful too.