Add files via upload
this file it's also considering the whole dataset and I arrived at the word frequency, where I identified some words that could be eliminated or subjects that I could eliminate the whole row. You will see that there are some signs ( " ", ...) that the code could not eliminate :( 
Also it is counting the blank spaces, but I dont think it is that huge problem.